pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
authors = ["AnastasiaTimoshe <a.n.timoshevskaya@gmail.com>"]
readme = "README.md"
packages = [
    { include = "task_manager" },
    ]

[tool.poetry.dependencies]
python = "^3.11"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.1"
django = "^5.0.7"
gunicorn = "21.2.0"
django-bootstrap5 = "^24.2"
uvicorn = "^0.29.0"
whitenoise = "^6.6.0"
django-filter = "^24.2"
rollbar = "^0.16.3"
coverage = "^7.6.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
